mobile: splitting out java utils from cronet
