+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/AndroidCertVerifyResult.java
@@ -1,4 +1,4 @@
-package org.chromium.net;
+package io.envoyproxy.envoymobile.utilities;
 
 import java.security.cert.CertificateEncodingException;
 import java.security.cert.X509Certificate;
