+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/AndroidNetworkLibrary.java
@@ -1,4 +1,4 @@
-package org.chromium.net;
+package io.envoyproxy.envoymobile.utilities;
 
 import android.net.TrafficStats;
 import android.os.ParcelFileDescriptor;
@@ -92,6 +92,7 @@ public final class AndroidNetworkLibrary {
    */
   public static void addTestRootCertificate(byte[] rootCert)
       throws CertificateException, KeyStoreException, NoSuchAlgorithmException {
+    System.err.println("HERE MADE THE FUNCTION");
     if (mUseFakeCertificateVerification) {
       FakeX509Util.addTestRootCertificate(rootCert);
     } else {
