+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/ContextUtils.java
@@ -1,4 +1,4 @@
-package org.chromium.net;
+package io.envoyproxy.envoymobile.utilities;
 
 import android.app.Activity;
 import android.app.Application;
