+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/StrictModeContext.java
@@ -1,4 +1,4 @@
-package org.chromium.net;
+package io.envoyproxy.envoymobile.utilities;
 
 import android.os.Build;
 import android.os.StrictMode;
