+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/ThreadStatsUid.java
@@ -1,4 +1,4 @@
-package org.chromium.net.impl;
+package io.envoyproxy.envoymobile.utilities;
 
 import android.net.TrafficStats;
 import java.lang.reflect.InvocationTargetException;
