+++ mobile/library/java/io/envoyproxy/envoymobile/utilities/UtilitiesManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="io.envoyproxy.envoymobile.utilities">
+    <uses-sdk
+            android:minSdkVersion="21"
+            android:targetSdkVersion="29"/>
+
+</manifest>
