+++ mobile/library/java/org/chromium/net/impl/BUILD
@@ -41,7 +41,6 @@ android_library(
         "Preconditions.java",
         "QuicExceptionImpl.java",
         "RequestFinishedInfoImpl.java",
-        "ThreadStatsUid.java",
         "UrlRequestBase.java",
         "UrlRequestBuilderImpl.java",
         "UrlResponseInfoImpl.java",
@@ -54,6 +53,7 @@ android_library(
         "//library/java/io/envoyproxy/envoymobile/engine:envoy_base_engine_lib",
         "//library/java/io/envoyproxy/envoymobile/engine:envoy_engine_lib",
         "//library/java/io/envoyproxy/envoymobile/engine/types:envoy_c_types_lib",
+        "//library/java/io/envoyproxy/envoymobile/utilities",
         "//library/java/org/chromium/net",
         "//library/java/org/chromium/net/urlconnection",
         artifact("androidx.annotation:annotation"),
