+++ mobile/test/java/io/envoyproxy/envoymobile/utilities/CertificateVerificationTest.java
@@ -1,4 +1,4 @@
-package org.chromium.net;
+package io.envoyproxy.envoymobile.utilities;
 
 import static org.junit.Assert.assertEquals;
