+++ mobile/test/java/org/chromium/net/BUILD
@@ -37,22 +37,6 @@ envoy_mobile_android_test(
 )
 
 envoy_mobile_android_test(
-    name = "certificate_verification_tests",
-    srcs = [
-        "CertificateVerificationTest.java",
-    ],
-    native_deps = [
-        "//test/common/jni:libenvoy_jni_with_test_extensions.so",
-        "//test/common/jni:libenvoy_jni_with_test_extensions_jnilib",
-    ],
-    deps = [
-        "//library/java/io/envoyproxy/envoymobile/engine:envoy_base_engine_lib",
-        "//library/java/io/envoyproxy/envoymobile/engine:envoy_engine_lib",
-        "//library/java/org/chromium/net",
-    ],
-)
-
-envoy_mobile_android_test(
     name = "cronet_url_request_context_test",
     srcs = [
         "CronetUrlRequestContextTest.java",
